resume_verifier/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
awsebcli==3.21.0
blessed==1.20.0
botocore==1.35.99
cachetools==5.5.0
cement==2.10.14
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
frozenlist==1.5.0
google-auth==2.37.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
<<<<<<< HEAD
jmespath==1.0.1
multidict==6.1.0
numpy==2.2.1
openai==1.12.0
pathspec==0.10.1
pdfminer.six==20240706
playwright==1.49.1
propcache==0.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
=======
google-generativeai==0.8.3
>>>>>>> a874d34 (Replace OpenAI/Claude APIs with Gemini API and fix various issues)
pydantic==2.10.4
pydantic_core==2.27.2
pyee==12.0.0
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9
semantic-version==2.10.0
scikit-learn==1.6.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
termcolor==2.5.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
yarl==1.18.3
